Improved Booster Tarots v1
Improved Booster Tarots or IBT is a retexture mod for Balatro

It replaces the textures of all the "booster" tarots (e.g. tarots that change your playing cards) to make them look like their effects.

Credit to IAmMe for making the Empress sprite and for the ideas for the Hierophant and Lovers, they're the reason I made this mod.